fix: AI 审查修复批次——限价合并 Int 溢出封顶 + aspects 服务端上限 64（求购单/黑名单/价格限制）+ 物品搜索 itemIds 钳制+Set 化 + 拍卖创建防重 + 出价弹窗 BID 广播不覆盖输入 + 购买总价 Long 计算 + 管理端翻页按钮 inFlight 置灰修复 + 封禁理由长度钳制

+++ common/src/client/kotlin/com/shusheng/cobblemarket/screen/ItemMarketScreen.kt
@@ -669,10 +669,11 @@ class ItemMarketScreen : Screen(Text.translatable("cobblemarket.item.title")) {
         closeBuyDialog()
     }
 
-    private fun buyTotal(): Int {
+    /** 总价 Long 计算：单价 9 位 × 数量 4 位可超 Int 上限，Int 相乘会溢出显示负数 */
+    private fun buyTotal(): Long {
         val entry = selectedEntry ?: return 0
         val count = buyCountField?.text?.toIntOrNull() ?: 0
-        return entry.price * count
+        return entry.price.toLong() * count
     }
 
     private fun renderBuyDialogBackground(context: DrawContext) {
